refactor bot-request-utils
