+++ packages/experiments-realm/commands/create-show-card-request.ts
@@ -3,8 +3,11 @@ import { Command } from '@cardstack/runtime-common';
 import { CreateShowCardRequestInput } from 'https://cardstack.com/base/command';
 
 import UseAiAssistantCommand from '@cardstack/boxel-host/commands/ai-assistant';
-import InviteUserToRoomCommand from '@cardstack/boxel-host/commands/invite-user-to-room';
 import SendBotTriggerEventCommand from '@cardstack/boxel-host/commands/send-bot-trigger-event';
+import {
+  ensureSubmissionBotIsInRoom,
+  realmURLFromCardId,
+} from './bot-request-utils';
 
 export default class CreateShowCardRequestCommand extends Command<
   typeof CreateShowCardRequestInput,
@@ -46,43 +49,4 @@ export default class CreateShowCardRequestCommand extends Command<
       },
     });
   }
-}
-
-async function ensureSubmissionBotIsInRoom(
-  commandContext: CreateShowCardRequestCommand['commandContext'],
-  roomId: string,
-) {
-  try {
-    await new InviteUserToRoomCommand(commandContext).execute({
-      roomId,
-      userId: 'submissionbot',
-    });
-  } catch (error) {
-    if (
-      !(error instanceof Error) ||
-      !error.message.includes('user already in room')
-    ) {
-      throw error;
-    }
-  }
-}
-
-function realmURLFromCardId(cardId: string): string {
-  let url: URL;
-  try {
-    url = new URL(cardId);
-  } catch {
-    throw new Error('cardId must be an absolute URL');
-  }
-
-  let pathSegments = url.pathname.split('/').filter(Boolean);
-  if (pathSegments.length < 2) {
-    throw new Error('cardId must include card type and card slug');
-  }
-
-  let realmSegments = pathSegments.slice(0, -2);
-  url.pathname = `/${realmSegments.join('/')}${realmSegments.length ? '/' : ''}`;
-  url.search = '';
-  url.hash = '';
-  return url.href;
 }
